FreeBSD 11 stable 6ebace841cdae274f1dcecf4eefa6b9cc2a8a469 May 2020, around 11.4 release.

+++ sbin/restore/extern.h
@@ -27,7 +27,7 @@
  * SUCH DAMAGE.
  *
  *	@(#)extern.h	8.2 (Berkeley) 1/7/94
- * $FreeBSD: release/10.0.0/sbin/restore/extern.h 203155 2010-01-29 10:00:42Z jh $
+ * $FreeBSD$
  */
 
 struct entry	*addentry(char *, ino_t, int);
@@ -54,8 +54,8 @@ void		 freeentry(struct entry *);
 void		 freename(char *);
 int	 	 genliteraldir(char *, ino_t);
 char		*gentempname(struct entry *);
-void		 getfile(void (*)(char *, long), void (*)(char *, long),
-			void (*)(char *, long));
+void		 getfile(void (*)(char *, size_t), void (*)(char *, size_t),
+			void (*)(char *, size_t));
 void		 getvol(long);
 void		 initsymtable(char *);
 int	 	 inodetype(ino_t);
@@ -98,7 +98,7 @@ void		 swabst(u_char *, u_char *);
 void	 	 treescan(char *, ino_t, long (*)(char *, ino_t, int));
 ino_t		 upperbnd(ino_t);
 long		 verifyfile(char *, ino_t, int);
-void		 xtrnull(char *, long);
+void		 xtrnull(char *, size_t);
 
 /* From ../dump/dumprmt.c */
 void		rmtclose(void);
